% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{innerprod}
\alias{innerprod}
\title{inner product}
\usage{
innerprod(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\description{
this function computes an inner product of two vectors.
}
