% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pick_node}
\alias{pick_node}
\title{pick a node}
\usage{
pick_node(wt, cut_pnt, options)
}
\arguments{
\item{wt}{a numeric scalar}

\item{cut_pnt}{a numeric scalar}

\item{options}{a character vector of options}
}
\description{
this function determines which node new observations go to
}
