% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblique_survival_tree_predict.R
\name{predict.ost}
\alias{predict.ost}
\title{Compute predictions using an oblique survival tree}
\usage{
\method{predict}{ost}(object, newdata, times, internal = FALSE, ...)
}
\arguments{
\item{object}{The oblique survival tree used to compute predictions.}

\item{newdata}{A data frame containing observations to predict.}

\item{times}{A vector of times in the range of the response variable, e.g. times when the response is a survival object, at which to return the survival probabilities.}

\item{internal}{This argument is only used when computing out of bag error for oblique random survival forests and does not require manual supervision.}

\item{...}{Other arguments passed to or from other functions.}
}
\value{
A matrix of survival probabilities containing 1 row for each observation and 1 column for each value in times.
}
\description{
Compute predictions using an oblique survival tree
}
\examples{

data("pbc",package='survival')
pbc$status[pbc$status>=1]=pbc$status[pbc$status>=1]-1
pbc$id=NULL
fctrs<-c('trt','ascites','spiders','edema','hepato','stage')
for(f in fctrs)pbc[[f]]=as.factor(pbc[[f]])
pbc=na.omit(pbc)

ost=OST(data=pbc, minsplit = 125)

#One and 2 year survival probability
times=seq(365, 365*2,length.out = 250)

prds=predict(ost,newdata=pbc,times=times)
}
