% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_orsf}
\alias{predict_orsf}
\title{Predictions from an ORSF}
\usage{
predict_orsf(forest, newx, times)
}
\arguments{
\item{forest}{A list of oblique survival trees.}

\item{newx}{a data matrix.}

\item{times}{A vector of times in the range of the response variable, e.g. times when the response is a survival object, at which to return the survival probabilities.}
}
\value{
A matrix of survival probabilities containing 1 row for each observation and 1 column for each value in times.
}
\description{
Predictions from an ORSF
}
