% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_ost}
\alias{predict_ost}
\title{Predictions from an OST}
\usage{
predict_ost(ostree, newx, times)
}
\arguments{
\item{ostree}{an oblique survival tree object.}

\item{newx}{a data matrix.}

\item{times}{A vector of times in the range of the response variable, e.g. times when the response is a survival object, at which to return the survival probabilities.}
}
\value{
A matrix of survival probabilities containing 1 row for each observation and 1 column for each value in times.
}
\description{
Predictions from an OST
}
