% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblique_survival_forest_predictSurvProb.R,
%   R/predictSurvProb_orsf.R
\name{predictSurvProb.orsf}
\alias{predictSurvProb.orsf}
\title{Compute predictions using an oblique random survival forest.}
\usage{
\method{predictSurvProb}{orsf}(object, newdata, times, ...)

\method{predictSurvProb}{orsf}(object, newdata, times, ...)
}
\arguments{
\item{object}{An object fitted using the ORSF function.}

\item{newdata}{A data frame containing observations to predict.}

\item{times}{A vector of times in the range of the response variable, e.g. times when the response is a survival object, at which to return the survival probabilities.}

\item{...}{Additional parameters for predict.ost}
}
\value{
A matrix of survival probabilities containing 1 row for each observation and 1 column for each value in times.

A matrix of survival probabilities containing 1 row for each observation and 1 column for each value in times.
}
\description{
Compute predictions using an oblique random survival forest.

Compute predictions using an oblique random survival forest.
}
\examples{

data("pbc",package='survival')
pbc$status[pbc$status>=1]=pbc$status[pbc$status>=1]-1
pbc$id=NULL
fctrs<-c('trt','ascites','spiders','edema','hepato','stage')
for(f in fctrs)pbc[[f]]=as.factor(pbc[[f]])
pbc=na.omit(pbc)

orsf=ORSF(data=pbc,ntree=10,minsplit=75)
times=seq(365, 365*2,length.out = 250)

prds=pec::predictSurvProb(orsf,newdata=pbc,times=times)
}
