% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{surv_est}
\alias{surv_est}
\title{Survival estimate}
\usage{
surv_est(times, probs, newtimes)
}
\arguments{
\item{times}{a numeric vector}

\item{probs}{a numeric vector}

\item{newtimes}{a numeric vector}
}
\description{
this function is used to create survival step estimates.
}
