% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Plot variable dependence using an oblique random survival forest}
\usage{
theme_Publication(base_size = 16)
}
\arguments{
\item{base_size}{how big to make the text}
}
\description{
Plot variable dependence using an oblique random survival forest
}
