% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{which_trees}
\alias{which_trees}
\title{which trees}
\usage{
which_trees(vec, forest)
}
\arguments{
\item{vec}{a numeric vector}

\item{forest}{a list of oblique survival trees}
}
\description{
this function identifies out-of-bag trees.
}
