% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{obliqueRSF}
\alias{obliqueRSF}
\alias{obliqueRSF-package}
\title{Oblique Random Survival Forests}
\description{
Oblique random survival forest (ORSFs) are ensembles for right-censured survival data that use linear combinations of input variables to recursively partition a set of training data. Regularized Cox proportional hazard models identify optimal linear combinations of input variables in each recursive partitioning step while building survival trees.
}
\author{
Byron C. Jaeger <bcjaeger@uab.edu>
}
