% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_dependence_plot.R
\name{vdplot}
\alias{vdplot}
\title{Plot variable dependence using an oblique random survival forest}
\usage{
vdplot(object, xvar, include.hist = TRUE, nonevent_lab = "survival",
  fvar = NULL, flab = NULL, time_units = "years", xlab = xvar,
  xvar_units = NULL, xlvls = NULL, sub_times = NULL)
}
\arguments{
\item{object}{an ORSF object (i.e. object returned from the ORSF function)}

\item{xvar}{a string giving the name of the x-axis variable}

\item{include.hist}{if true, a histogram showing the distribution of values for the x-axis variable will be included at the bottom of the plot.}

\item{nonevent_lab}{the label that describes a non-event.}

\item{fvar}{(optional) a string indicating a variable to facet the plot with}

\item{flab}{the labels to be printed describing the facet variable. For a facet variable with k categories, flab should be a vector with k labels, given in the same order as the levels of the facet variable.}

\item{time_units}{the unit of time, e.g. days, since baseline.}

\item{xlab}{the label to be printed describing the x-axis variable}

\item{xvar_units}{the unit of measurement for the x-axis variable. For example, age is usually measured in years.}

\item{sub_times}{sometimes you just want}
}
\value{
A ggplot2 object
}
\description{
Plot variable dependence using an oblique random survival forest
}
\examples{

data("pbc",package='survival')
pbc$status[pbc$status>=1]=pbc$status[pbc$status>=1]-1
pbc$time=pbc$time/365.25
pbc$id=NULL
fctrs<-c('trt','ascites','spiders','edema','hepato','stage')
for(f in fctrs)pbc[[f]]=as.factor(pbc[[f]])
pbc=na.omit(pbc)

orsf=ORSF(data=pbc, eval_time=5)

vdplot(object=orsf, xvar='bili', xlab='Bilirubin', xvar_units='mg/dl')

}
